+++ lib/extensions.dart
@@ -1,0 +1,3 @@
+extension Precision on double {
+  double toPrecision(int n) => double.parse(toStringAsFixed(n));
+}