+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## [2.7.2]
+- Mini fix and cleanup
+
+## [2.7.1]
+- Improved docs
+- Increased the default maxDigits parameter to 15
+
+## [2.7.0]
+- Removed decimal dependency in favor of int getter
+- Improved sample with the new int initializer and getter
+
+## [2.5.2]
+- Improved sample with custom input field
+
+## [2.5.1]
+- Bump decimal and dart versions
+- Added missing type of some variables, different getters of the controller on example and linter on code
+
+## [2.5.0]
+- Fixed cursor position (tks @benz93chung)
+- Fixed double value not reseting when clearing content from text field
+- added init value
+- added precision to double value
+
 ## [2.0.0]
 - Migrate to null-safety
 ## [1.0.1] - [1.0.2] - [1.0.3] - 2019-10-07.